--- client/src/types/result.ts
@@ -1,3 +1,0 @@
-export type Result<T, E = Error> =
-  | { ok: true; value: T }
-  | { ok: false; error: E };
